test(rule): unify testdata into scenario dirs run by test_testdata
Fold the edit-scenario runner into test_testdata.py so a single test drives
the full lifecycle of every scenario under testdata/resources/signoz_<name>/:
create (plan shows a create -> apply -> re-plan clean), then each
NN-jsonpatch.json edit (re-plan shows changes -> apply -> re-plan clean), then
destroy. A patch-free scenario is just create -> no-drift -> destroy.

Every config is now a numbered scenario directory with an 01-<name>.tf(.json)
base; the flat *.tf edge cases move into 02-05. Remove test_editing.py (its
logic now lives in test_testdata.py) and refresh the testdata README.
