+++ tests/testdata/resources/signoz_rule/02/01-anomaly_seasonality.tf
