+++ tests/testdata/resources/signoz_rule/03/01-clickhouse_query.tf
